requirements.txt
aws-cdk-lib==2.163.1
constructs>=10.0.0,<11.0.0
delta-spark==2.3.0
